bug test calculateAvgGradeOfClassworkAssignmentInCourse.returns an array of data
