+++ packages/server/src/modules/academic/academic.module.ts
@@ -9,19 +9,17 @@ import { OrgOfficeModule } from 'modules/orgOffice/orgOffice.module'
 
 import { AcademicService } from './academic.service'
 import { AcademicSubjectResolver } from './academicSubject.resolver'
-import { CourseResolver } from './course.resolver'
 import { AcademicSubject } from './models/AcademicSubject'
-import { Course } from './models/Course'
 
 @Module({
   imports: [
-    TypegooseModule.forFeature([AcademicSubject, Course, ClassworkAssignment]),
+    TypegooseModule.forFeature([AcademicSubject, ClassworkAssignment]),
     forwardRef(() => OrgModule),
     forwardRef(() => OrgOfficeModule),
     forwardRef(() => AccountModule),
     forwardRef(() => FileStorageModule),
   ],
-  providers: [AcademicSubjectResolver, AcademicService, CourseResolver],
+  providers: [AcademicSubjectResolver, AcademicService],
   exports: [AcademicService],
 })
 export class AcademicModule {}
