+++ packages/server/src/modules/auth/auth.module.ts
@@ -2,6 +2,7 @@ import { forwardRef, Global, Module } from '@nestjs/common'
 
 import { AcademicModule } from 'modules/academic/academic.module'
 import { AccountModule } from 'modules/account/account.module'
+import { CourseModule } from 'modules/course/course.module'
 import { OrgModule } from 'modules/org/org.module'
 
 import { AuthResolver } from './auth.resolver'
@@ -12,7 +13,8 @@ import { AuthService } from './auth.service'
   imports: [
     forwardRef(() => AccountModule),
     forwardRef(() => AcademicModule),
-    OrgModule,
+    forwardRef(() => OrgModule),
+    forwardRef(() => CourseModule),
   ],
   providers: [AuthResolver, AuthService],
   exports: [AuthService],
