+++ packages/server/src/modules/classwork/classwork.module.ts
@@ -1,10 +1,10 @@
-import { Global, Module } from '@nestjs/common'
+import { forwardRef, Global, Module } from '@nestjs/common'
 import { TypegooseModule } from 'nestjs-typegoose'
 
-import { Course } from 'modules/academic/models/Course'
 import { AccountModule } from 'modules/account/account.module'
 import { Account } from 'modules/account/models/Account'
 import { AuthModule } from 'modules/auth/auth.module'
+import { Course } from 'modules/course/models/Course'
 import { FileStorageModule } from 'modules/fileStorage/fileStorage.module'
 import { MailModule } from 'modules/mail/mail.module'
 import { OrgModule } from 'modules/org/org.module'
@@ -20,11 +20,6 @@ import { ClassworkSubmission } from './models/ClassworkSubmission'
 @Global()
 @Module({
   imports: [
-    AuthModule,
-    AccountModule,
-    FileStorageModule,
-    OrgModule,
-    MailModule,
     TypegooseModule.forFeature([
       ClassworkAssignment,
       ClassworkMaterial,
@@ -32,6 +27,11 @@ import { ClassworkSubmission } from './models/ClassworkSubmission'
       Course,
       Account,
     ]),
+    forwardRef(() => AuthModule),
+    forwardRef(() => AccountModule),
+    forwardRef(() => FileStorageModule),
+    forwardRef(() => OrgModule),
+    forwardRef(() => MailModule),
   ],
   providers: [
     ClassworkService,
