+++ packages/server/src/modules/course/models/Course.ts
