+++ packages/server/src/modules/lesson/lesson.module.ts
@@ -2,7 +2,7 @@ import { forwardRef, Global, Module } from '@nestjs/common'
 import { TypegooseModule } from 'nestjs-typegoose'
 
 import { AcademicModule } from 'modules/academic/academic.module'
-import { Course } from 'modules/academic/models/Course'
+import { Course } from 'modules/course/models/Course'
 
 import { LessonResolver } from './lesson.resolver'
 import { LessonService } from './lesson.service'
